chore: move types into shared folder
this way different layers explicitly can access these types
