+++ src/adapters/sqlalchemy_repository.py
@@ -4,10 +4,8 @@ from typing import List, Optional
 from sqlalchemy.orm import Session
 
 from src.adapters.repository import AbstractRepository
-from src.domain.model import (
-    UUID,
-    Task,
-)
+from src.domain.model import Task
+from src.shared.types import UUID
 
 
 class SQLAlchemyRepository(AbstractRepository):
