+++ src/shared/types.py
@@ -1,0 +1,23 @@
+from enum import StrEnum
+from typing import NewType
+
+UUID = NewType("UUID", str)
+
+
+class TaskType(StrEnum):
+    SCRIPT = "script"
+    MODEL = "model"
+    UNKNOWN = "unknown"
+
+
+class TaskStatus(StrEnum):
+    PENDING = "pending"
+    RUNNING = "running"
+    COMPLETED = "completed"
+    FAILED = "failed"
+    UNKNOWN = "unknown"
+
+
+class Model(StrEnum):
+    VTC = "vtc"
+    UNKNOWN = "unknown"
